refactor(approximators): move traits inside module and rename FA types
